requirements.txt
asgiref==3.2.7
autopep8==1.5.4
Brotli==1.0.9
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.7
django-cacheops==5.0.1
django-cleanup==5.0.0
django-cors-headers==3.3.0
django-debug-toolbar==2.2
django-heroku==0.3.1
django-querycount==0.7.0
django-rest-knox==4.1.0
django-storages==1.9.1
djangorestframework==3.11.0
environs==8.0.0
funcy==1.14
google-api-core==1.22.1
google-auth==1.20.1
google-cloud-core==1.4.1
google-cloud-storage==1.30.0
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
idna==2.10
marshmallow==3.6.1
mypy==0.782
mypy-extensions==0.4.3
Pillow==7.2.0
protobuf==3.13.0
psycopg2==2.8.6
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
python-dotenv==0.13.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
rsa==4.6
sentry-sdk==0.16.5
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.10
whitenoise==5.2.0
